+++ service/src/main/java/by/work/service/OrderServiceImpl.java
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.sql.Date;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -48,7 +47,7 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public void update(Order order, Double totalAmount) {
+    public void update(Order order, BigDecimal totalAmount) {
         order.setTotalAmount(totalAmount);
         orderRepository.save(order);
     }
